Set up the navigation for onboarding screen
